+++ src/components/ui/ImageGallery/ImageGallerySelector.tsx
@@ -36,6 +36,7 @@ const hasScroll = (container: HTMLDivElement | null): boolean => {
 
 function ImageGallerySelector({ images, onSelect, currentImageIdx }: Props) {
   const elementsRef = useRef<HTMLDivElement>(null)
+  const elementHasScroll = hasScroll(elementsRef.current)
   const { ref: firstImageRef, inView: firstImageInView } = useInView({
     threshold: 1,
   })
@@ -51,7 +52,7 @@ function ImageGallerySelector({ images, onSelect, currentImageIdx }: Props) {
       aria-roledescription="carousel"
       aria-label="Product images"
     >
-      {hasScroll(elementsRef.current) && !firstImageInView && (
+      {elementHasScroll && !firstImageInView && (
         <IconButton
           aria-label="backward slide image selector"
           icon={<Icon name="ArrowLeft" width={24} height={24} />}
@@ -60,13 +61,12 @@ function ImageGallerySelector({ images, onSelect, currentImageIdx }: Props) {
       )}
       <div data-fs-image-gallery-selector-elements ref={elementsRef}>
         {images.map((image, idx) => {
-          let ref
-
-          if (idx === 0) {
-            ref = firstImageRef
-          } else if (idx === images.length - 1) {
-            ref = lastImageRef
-          }
+          const ref =
+            idx === 0
+              ? firstImageRef
+              : idx === images.length - 1
+              ? lastImageRef
+              : null
 
           return (
             <Button
@@ -92,7 +92,7 @@ function ImageGallerySelector({ images, onSelect, currentImageIdx }: Props) {
           )
         })}
       </div>
-      {hasScroll(elementsRef.current) && !lastImageInView && (
+      {elementHasScroll && !lastImageInView && (
         <IconButton
           aria-label="forward slide image selector"
           icon={<Icon name="ArrowLeft" width={24} height={24} />}